Reforzar API y persistencia del backend

+++ backend/models.py
@@ -1,5 +1,5 @@
 from database import Base
-from sqlalchemy import Column, Date, Float, Integer, String
+from sqlalchemy import Column, Date, Integer, Numeric, String
 
 
 class Transaccion(Base):
@@ -7,7 +7,7 @@ class Transaccion(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     titulo = Column(String, index=True, nullable=True)
-    monto = Column(Float, nullable=False)
+    monto = Column(Numeric(12, 2), nullable=False)
     tipo = Column(String, index=False, nullable=False)
     categoria = Column(String, index=True, nullable=True)
     fecha = Column(Date, nullable=False)
